trivial: Split __init__.py into smaller files
