+++ wikipediaapi/_version.py
@@ -1,0 +1,4 @@
+"""Version information for Wikipedia-API."""
+
+__version__ = (0, 9, 0)
+__version_str__ = ".".join(str(s) for s in __version__)
